+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `Innmind\IO\Files\Temporary::close()`
+- `Innmind\IO\Files\Temporary::push()`
 
 ## 3.1.0 - 2025-04-21